Fixed a "few" race conditions in the tests
Looked like there was a decent amount of unsynchronized access of
data in the tests. Typical things included declaring arrays on
the main thread and appending to them in a callback (that's not
guaranteed to land on the main thread)

And then the Testing Networking Proxy also had a slight bug that
stored the call data after the completion callback was given to
the client. So I divided the calldata in to request data that is
guaranteed to be there before the client completion block is called
but after the networking proxy actually makes the call, and a
responses object which is separate an only available upon return
of the callback