Remove EventEmitter and fix AutoRefreshingProtocol
For some silly reason, a task manager was for a user inside
the protocol and stored in a map, but the idiot who wrote it
didn't realize that user already has a taskManager so there's
no reason to do that.

Also the executor in the TaskOperation had a redundant done
callback that just set isFinished to true immdiately after
going in the execution block. So removed that